requirements.txt
numpy==1.18.3
tf-nightly-gpu==2.3.0.dev20200530
tf-agents-nightly==0.5.0.dev20200530